Added percentages and fractions to custom numTab
This commit is a minor addition to the numtab function, that does not interfere with any existing functionality. It adds the small feature of allowing players to input fractions and percentages into the custom field. These are then interpreted as fractions of the total number of unemployed trimps. This can, for example, allow a player to enter 1/3 (or 33% respectively) and then distribute his unemployed trimps equally on Farmer, Lumber and Miner. The tooltipTexxt for the custom tab has been expanded to reflect that new feature.