AutoPants fix
The secret AutoPrestige setting, PANTS ONLY, would actually buy any prestige
as long as all available prestiges belonged to the same category (all armors
or all weapons).

This was due to some logic shared with Weapons First / Prestige All. The fix
was to move the PANTS ONLY check earlier, in line with what was done for
Weapons Only.