HTML validation (fixes #38)
The game’s HTML contained some constructs that are invalid according to the
spec. The fixes are pretty straightforward.

This second difficultyAdvMapsText div under the Biome section really looked like
a copy-paste mistake, so I deleted it outright instead of simply changing the
ID. It didn’t cause any problems in my tests.