Don’t signal turkimp buff when it’s buffing 0 workers
When gathering food, wood or metal with a turkimp buff, the gathering icon
changes color and gains a `spoon-knife` icon, signalling that the workers
assigned to this resource are buffed by sharing food.

However, this can be misleading. When no workers are assigned to the resource in
question, the buff has very little effect (+50% of 0 production is still 0; the
+16.9% loot still applies, but is usually minor).

This commit adds a check so that the color and icon are only applied if workers
are assigned to the resource gathered by the player. This is also updated when
hiring or firing trimps.